src/wingsight-server/requirements.txt
asgiref==3.8.1
boto3==1.37.34
botocore==1.37.34
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
cryptography==44.0.2
Django==5.1.6
django-cors-headers==4.7.0
djangorestframework==3.15.2
drf-yasg==1.21.10
idna==3.10
inflection==0.5.1
jmespath==1.0.1
numpy==2.2.4
packaging==24.2
pika==1.3.2
pycparser==2.22
PyJWT==2.10.1
PyMySQL==1.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.1
PyYAML==6.0.2
requests==2.32.3
s3transfer==0.11.4
six==1.17.0
sqlparse==0.5.3
uritemplate==4.1.1
urllib3==2.4.0
